+++ src/main/java/dinkplugin/DinkPlugin.java
@@ -25,13 +25,17 @@ import net.runelite.api.events.UsernameChanged;
 import net.runelite.api.events.VarbitChanged;
 import net.runelite.api.events.WidgetLoaded;
 import net.runelite.client.RuneLite;
+import net.runelite.client.chat.ChatMessageManager;
+import net.runelite.client.chat.QueuedMessage;
 import net.runelite.client.config.ConfigManager;
 import net.runelite.client.eventbus.Subscribe;
+import net.runelite.client.events.ConfigChanged;
 import net.runelite.client.events.NpcLootReceived;
 import net.runelite.client.events.PlayerLootReceived;
 import net.runelite.client.plugins.Plugin;
 import net.runelite.client.plugins.PluginDescriptor;
 import net.runelite.client.plugins.loottracker.LootReceived;
+import net.runelite.client.util.ColorUtil;
 import net.runelite.client.util.Text;
 
 import javax.inject.Inject;
@@ -40,10 +44,15 @@ import javax.inject.Inject;
 @PluginDescriptor(
     name = "Dink",
     description = "A notifier for sending webhooks to Discord or other custom destinations",
-    tags = { "loot", "logger", "collection", "pet", "death", "xp", "level", "notifications", "discord", "speedrun" }
+    tags = { "loot", "logger", "collection", "pet", "death", "xp", "level", "notifications", "discord", "speedrun",
+        "diary", "combat achievements", "combat task" }
 )
 public class DinkPlugin extends Plugin {
     public static final String USER_AGENT = RuneLite.USER_AGENT + " (Dink/1.x)";
+
+    private @Inject ChatMessageManager chatManager;
+
+    private @Inject SettingsValidator validator;
 
     private @Inject CollectionNotifier collectionNotifier;
     private @Inject PetNotifier petNotifier;
@@ -76,6 +85,11 @@ public class DinkPlugin extends Plugin {
     @Provides
     DinkPluginConfig provideConfig(ConfigManager configManager) {
         return configManager.getConfig(DinkPluginConfig.class);
+    }
+
+    @Subscribe
+    public void onConfigChanged(ConfigChanged event) {
+        validator.onConfigChanged(event);
     }
 
     @Subscribe
@@ -144,6 +158,7 @@ public class DinkPlugin extends Plugin {
 
     @Subscribe
     public void onVarbitChanged(VarbitChanged event) {
+        validator.onVarbitChanged(event);
         diaryNotifier.onVarbitChanged(event);
     }
 
@@ -152,5 +167,20 @@ public class DinkPlugin extends Plugin {
         questNotifier.onWidgetLoaded(event);
         clueNotifier.onWidgetLoaded(event);
         speedrunNotifier.onWidgetLoaded(event);
+    }
+
+    public void addChatWarning(String message) {
+        String formatted = String.format("[%s] %s: %s",
+            ColorUtil.wrapWithColorTag(getName(), Utils.PINK),
+            "Warning",
+            ColorUtil.wrapWithColorTag(message, Utils.RED)
+        );
+
+        chatManager.queue(
+            QueuedMessage.builder()
+                .type(ChatMessageType.CONSOLE)
+                .runeLiteFormattedMessage(formatted)
+                .build()
+        );
     }
 }