+++ src/test/java/dinkplugin/notifiers/DeathNotifierTest.java
@@ -57,6 +57,7 @@ class DeathNotifierTest extends MockedNotifierTest {
         when(config.notifyDeath()).thenReturn(true);
         when(config.deathNotifPvpEnabled()).thenReturn(true);
         when(config.deathSendImage()).thenReturn(false);
+        when(config.deathEmbedKeptItems()).thenReturn(true);
         when(config.deathNotifyMessage()).thenReturn("%USERNAME% has died, losing %VALUELOST% gp");
         when(config.deathNotifPvpMessage()).thenReturn("%USERNAME% has been PKed by %PKER% for %VALUELOST% gp");
 
@@ -119,8 +120,11 @@ class DeathNotifierTest extends MockedNotifierTest {
         List<SerializedItemStack> lost = Collections.singletonList(
             new SerializedItemStack(ItemID.TUNA, 1, TUNA_PRICE, "Tuna")
         );
-        List<NotificationBody.Embed> embeds = Collections.singletonList(
-            new NotificationBody.Embed(new NotificationBody.UrlEmbed("https://static.runelite.net/cache/item/icon/" + ItemID.TUNA + ".png"))
+        List<NotificationBody.Embed> embeds = Arrays.asList(
+            new NotificationBody.Embed(new NotificationBody.UrlEmbed("https://static.runelite.net/cache/item/icon/" + ItemID.RUBY + ".png")),
+            new NotificationBody.Embed(new NotificationBody.UrlEmbed("https://static.runelite.net/cache/item/icon/" + ItemID.SHARK + ".png")),
+            new NotificationBody.Embed(new NotificationBody.UrlEmbed("https://static.runelite.net/cache/item/icon/" + ItemID.OPAL + ".png")),
+            new NotificationBody.Embed(new NotificationBody.UrlEmbed("https://static.runelite.net/cache/item/icon/" + ItemID.COAL + ".png"))
         );
         verify(messageHandler).createMessage(
             PRIMARY_WEBHOOK_URL,